2.0.1

+++ buildall.sh
@@ -5,34 +5,26 @@ echo "****************************************************************"
 echo "*                        SCRIPT STARTED                        *"
 echo "****************************************************************"
 echo ""
-echo "******************* Updating from GitHub (DISCARDING ALL LOCAL CHANGES)..."
-git fetch --all
-git reset --hard origin/main
-git pull origin main
-echo ""
 
-echo "******************* Building cnxpack-forwarder..."
-cd cnxpack-forwarder
+if [[ $# -gt 0 ]] ; then
+    echo "******************* Updating from GitHub (DISCARDING ALL LOCAL CHANGES)..."
+    git fetch --all
+    git reset --hard origin/main
+    git pull origin main
+    echo ""
+fi
+
+echo "******************* Building forwarder..."
+cd app-forwarder
 make
 cd ..
 echo ""
 
-#echo "******************* Building cnxpack-rcm..."
-#cd cnxpack-rcm
-#make
-#cd ..
-#echo ""
-
-echo "******************* Building cnxpack-updater..."
+echo "******************* Building application..."
 make
 echo ""
 
-echo "******************* Creating the release ZIP..."
-mkdir switch
-mkdir ./switch/cnxpack-updater
-cp cnxpack-updater.nro ./switch/cnxpack-updater
-zip -r -9 ./cnxpack-updater.zip ./switch/
-rm -rf switch
+./release.sh
 
 echo ""
